Make it easier to break out tests into their own files
